Move EndpointSlice to io.fabric8.kubernetes.api.model.discovery.v1beta1
Moving EndpointSlice resource to apiVersion specific package name, I
future if this api gets promoted to v1, we would have lesser problems
adding it.

Related to #2702
