+++ kubernetes-model-generator/kubernetes-model-core/pom.xml
@@ -120,9 +120,9 @@
               io.fabric8.kubernetes.api.model.certificates.v1,
               io.fabric8.kubernetes.api.model.certificates.v1beta1,
               io.fabric8.kubernetes.api.model.coordination.v1,
-              io.fabric8.kubernetes.api.model.discovery,
               io.fabric8.kubernetes.api.model.events.v1,
               io.fabric8.kubernetes.api.model.events.v1beta1,
+              io.fabric8.kubernetes.api.model.discovery.v1beta1,
               io.fabric8.kubernetes.api.model.extensions,
               io.fabric8.kubernetes.api.model.metrics.v1beta1,
               io.fabric8.kubernetes.api.model.networking.v1,
