+++ kubernetes-model-generator/kubernetes-model-core/src/main/java/io/fabric8/kubernetes/internal/KubernetesDeserializer.java
@@ -150,9 +150,9 @@ public class KubernetesDeserializer extends JsonDeserializer<KubernetesResource>
                 "io.fabric8.kubernetes.api.model.certificates.v1beta1.",
                 "io.fabric8.kubernetes.api.model.coordination.v1.",
                 "io.fabric8.kubernetes.api.model.coordination.",
-                "io.fabric8.kubernetes.api.model.discovery.",
                 "io.fabric8.kubernetes.api.model.events.v1.",
                 "io.fabric8.kubernetes.api.model.events.v1beta1.",
+                "io.fabric8.kubernetes.api.model.discovery.v1beta1.",
                 "io.fabric8.kubernetes.api.model.metrics.v1beta1.",
                 "io.fabric8.kubernetes.api.model.networking.v1.",
                 "io.fabric8.kubernetes.api.model.networking.v1beta1.",
